requirements.txt
simple-salesforce==1.12.6
python-dotenv==1.1.1
scikit-learn==1.7.1
pandas==2.3.2
openpyxl==3.1.5